Feat: Product Changes component
