+++ src/components/navbar/index.js
@@ -1,5 +1,6 @@
 import React from "react";
 import {
+  GreenRangerPic,
   LeftArrow,
   MenuIcon,
   NotificationIcon,
@@ -17,22 +18,28 @@ const AdminNavbar = ({ name }) => {
       {name === "layout" ? (
         <div className="flex items-center gap-6">
           <MenuIcon alt="menu" className="w-[20px] h-[20px]" />
-          <p className="font-bold text-[20px] leading-[32px] text-[#186F3D]">
-            Green Ranger
-          </p>
+          <div className="flex">
+                <img src={GreenRangerPic} className="h-[32px] w-[32px]"/>
+              <p className="font-bold text-[20px] leading-[32px] text-[#186F3D] ml-2">
+                Green Ranger
+              </p>
+              </div>
         </div>
       ) : (
-        <div className="flex items-center gap-6">
+        <div className="flex items-center">
           {name === "viewOrders" ? (
-            <div>
+            <div className="flex items-center gap-6">
               <LeftArrow
                 alt="menu"
                 className="w-[20px] h-[20px] cursor-pointer"
                 onClick={() => navigate("/orders")}
               />
-              <p className="font-bold text-[20px] leading-[32px] text-[#186F3D]">
+              <div className="flex">
+                <img src={GreenRangerPic} className="h-[32px] w-[32px]"/>
+              <p className="font-bold text-[20px] leading-[32px] text-[#186F3D] ml-2">
                 All Stores
               </p>
+              </div>
             </div>
           ) : (
             <div className="flex items-center gap-[24px]">
@@ -41,11 +48,11 @@ const AdminNavbar = ({ name }) => {
                 className="w-[20px] h-[20px] cursor-pointer"
                 onClick={() => navigate("/products")}
               />
-              <div className="flex items-center gap-[8px]">
-                <ProfilePic className="w-[24px] h-[24px]" />
-                <p className="font-semibold text-[13px] leading-[23px] text-[#186F3D]">
-                  Ini James
-                </p>
+             <div className="flex">
+                <img src={GreenRangerPic} className="h-[32px] w-[32px]"/>
+              <p className="font-bold text-[20px] leading-[32px] text-[#186F3D] ml-2">
+                Green Ranger
+              </p>
               </div>
             </div>
           )}
@@ -54,7 +61,6 @@ const AdminNavbar = ({ name }) => {
 
       <div className="flex gap-4 items-center">
         <NotificationIcon className="w-[20px] h-[20px]" />
-        {name !== "AddProducts" && (
           <div  className="flex gap-4 items-center">
             <SettingsIcon className="w-[20px] h-[20px]" />
             <ProfilePic className="w-[24px] h-[24px]" />
@@ -62,7 +68,6 @@ const AdminNavbar = ({ name }) => {
               Ini James
             </p>
           </div>
-        )}
       </div>
     </nav>
   );
